backend/requirements.txt
Flask
requests
pyjwt
pymongo
psycopg2
python-dotenv
flask-cors
psycopg2-binary